Better deduplication strategy for querying in parallel
Before a simple `drop_duplicates` was used to deduplicate in parallel results. This led to results being dropped even if the index is different.

Moreover, when `store_keys` is True we expect a result will be added to each key in which it was found but if `store_keys` is False a unique result should only occur once in the output.

This commit implements this behavior.